cloudrun/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-cloud-storage==2.10.0
google-auth==2.23.4
google-api-python-client==2.108.0
httpx==0.25.1
pydantic[email]==2.5.0

